requirements.txt
Flask==3.0.0
flask-cors==4.0.0
mysql-connector-python==8.2.0
opencv-python==4.8.1.78
dlib==20.0.0
face-recognition==1.3.0
openpyxl==3.1.2
Pillow==10.1.0
numpy==1.24.3
werkzeug==3.0.1
python-dotenv==1.0.0
pyinstaller==6.3.0



